backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.340
langchain-openai==0.0.2
langchain-deeplake==0.1.1
deeplake==3.8.3
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
pydantic==2.5.0
